fix: share nextauth config and pass ci verify
