+++ app/page.tsx
@@ -1,5 +1,5 @@
+import { auth } from "@/auth";
 import Link from "next/link";
-import { auth } from "@/app/api/auth/[...nextauth]/route";
 import { redirect } from "next/navigation";
 
 export default async function HomePage() {
